refac exact solution algos
